backend/requirements.txt
# NOVA v3 Backend Dependencies
# FastAPI + Production Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Config
python-dotenv==1.0.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
psutil==5.9.8
